requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
langchain>=0.3.0,<0.4.0
langchain-text-splitters>=0.3.0,<0.4.0
langchain-openai>=0.2.0,<0.3.0
pydantic-settings>=2.5.0,<3.0.0
psycopg[binary]>=3.2.0,<4.0.0
pgvector>=0.3.0,<1.0.0
python-multipart>=0.0.9,<1.0.0
pypdf>=5.0.0,<6.0.0
python-docx>=1.1.0,<2.0.0
